+++ monkey/monkey_island/cc/services/config.py
@@ -2,6 +2,7 @@ import collections
 import copy
 import functools
 import logging
+from typing import Dict
 
 from jsonschema import Draft4Validator, validators
 
@@ -425,3 +426,20 @@ class ConfigService:
             ),
             "exploit_ssh_keys": ConfigService.get_config_value(SSH_KEYS_PATH, should_decrypt=False),
         }
+
+    @staticmethod
+    def format_flat_config_for_agent(config: Dict):
+        ConfigService._remove_credentials_from_flat_config(config)
+
+    @staticmethod
+    def _remove_credentials_from_flat_config(config: Dict):
+        fields_to_remove = {
+            "exploit_lm_hash_list",
+            "exploit_ntlm_hash_list",
+            "exploit_password_list",
+            "exploit_ssh_keys",
+            "exploit_user_list",
+        }
+
+        for field in fields_to_remove:
+            config.pop(field, None)