+++ CHANGELOG.md
@@ -32,6 +32,7 @@ Changelog](https://keepachangelog.com/en/1.0.0/).
 - Hostname system info collector. #1535
 - Max iterations and timeout between iterations config options. #1600
 - MITRE ATT&CK configuration screen. #1532
+- Propagation credentials from "GET /api/monkey/<string:guid>" endpoint. #1538
 
 ### Fixed
 - A bug in network map page that caused delay of telemetry log loading. #1545